Struct libcryptsetup_rs::CryptDevice

pub struct CryptDevice { /* fields omitted */ }

Data type that is a handle for a crypt device

Methods

impl CryptDevice

pub fn from_ptr(ptr: *mut crypt_device) -> Self

Reconstruct a CryptDevice object from a pointer

pub fn logging_handle(&mut self) -> CryptLog

Get a logging option handle

pub fn settings_handle(&mut self) -> CryptSettings

Get a settings option handle

pub fn format_handle(&mut self) -> CryptFormat

Get a format option handle

pub fn context_handle(&mut self) -> CryptContext

Get a context option handle

pub fn keyslot_handle(&mut self, keyslot: Option<c_int>) -> CryptKeyslot

Get a keyslot option handle

pub fn runtime_handle<'a>(&'a mut self, name: &'a str) -> CryptRuntime<'a>

Get a runtime attribute option handle

pub fn luks2_flag_handle<T>(&mut self) -> CryptLuks2Flags<T>

Get LUKS2 flags option handle

pub fn activate_handle(&mut self) -> CryptActivation

Get activation option handle

pub fn volume_key_handle(&mut self) -> CryptVolumeKey

Get volume key option handle

pub fn status_handle(&mut self) -> CryptDeviceStatus

Get crypt device status option handle

pub fn backup_handle(&mut self) -> CryptBackup

Get crypt device backup option handle

pub fn debug_handle() -> CryptDebug

Get crypt debug option handle

pub fn keyfile_handle(&mut self) -> CryptKeyfile

Get crypt device keyfile option handle

pub fn wipe_handle(&mut self) -> CryptWipe

Get crypt device wipe option handle

pub fn token_handle(&mut self, token: c_int) -> CryptLuks2Token

Get crypt device LUKS2 token option handle

pub fn reencrypt_handle(&mut self) -> CryptLuks2Reencrypt

Get crypt device reencryption option handle

pub fn set_confirm_callback<T>(
&mut self,
confirm: Option<unsafe extern "C" fn(msg: *const c_char, usrptr: *mut c_void) -> c_int>,
usrdata: Option<&mut T>
)

Set the callback that prompts the user to confirm an action

pub fn set_data_device(&mut self, device_path: &Path) -> Result<(), LibcryptErr>

Set the device path for a data device

pub fn set_data_offset(&mut self, offset: u64) -> Result<(), LibcryptErr>

Set the offset in 4096-byte sectors for the data section on a device