Trait libcryptsetup_rs::SafeMemzero

pub trait SafeMemzero {
    fn safe_memzero(&mut self);
}

A trait to be implemented for a segment of memory that can be explicitly zeroed in a way that will not be optimized away by the compiler.

Required methods

fn safe_memzero(&mut self)

Zero the data in the buffer. To enable managed zeroing of a buffer, call this in a Drop implementation.

Loading content...